Move actions up
